Restore gray coloring for autocomments
The main criticism of T165189 has been the "sea of blue" it causes. Given that
the most important part is the actual comment, other elements like the auto-
comment should be deemphasized (per Anomie).

This won't bring back exactly the same color scheme as before, as now the
arrow will also be colored gray instead of blue.

I don't think this is a permanent solution, and we should continue discussing
and iterating on how to address the prominence vs. accessibility issues.

It was also noted that the current gray is not fully accessible (T125657),
which should be addressed in a separate patch.

Bug: T165189
Change-Id: Ic8903a539017a3fd3d47627f8485caed09fce482
(cherry picked from commit c3cede1d738a0d57e05acdb776205e1f41d1e6af)